Move TypeMarshaller to interpreter. Make Marshal.Type a required param - just to be more explicit.
