+++ src/libraries/Microsoft.PowerFx.Interpreter/Marshal/PrimitiveMarshallerProvider.cs
@@ -1,0 +1,32 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using Microsoft.PowerFx.Core;
+using Microsoft.PowerFx.Core.Public.Types;
+using Microsoft.PowerFx.Core.Public.Values;
+
+namespace Microsoft.PowerFx
+{
+    /// <summary>
+    /// Marshalling provider to handle builtin primitive types. 
+    /// </summary>
+    public class PrimitiveMarshallerProvider : ITypeMarshallerProvider
+    {      
+        /// <inheritdoc/>
+        public bool TryGetMarshaller(Type type, TypeMarshallerCache cache, int maxDepth, out ITypeMarshaller marshaler)
+        {
+            if (BuiltinFormulaTypeConversions.TryGetFormulaType(type, out var fxType))
+            {
+                marshaler = new PrimitiveTypeMarshaler(fxType);
+                return true;
+            }
+
+            // Not supported
+            marshaler = null;
+            return false;
+        }
+    }
+}
