+++ src/tests/Microsoft.PowerFx.Core.Tests/PublicSurfaceTests.cs
@@ -26,15 +26,7 @@ namespace Microsoft.PowerFx.Core.Tests
                 "Microsoft.PowerFx.Core.Texl.Intellisense.SignatureHelp.ParameterInformation",
                 "Microsoft.PowerFx.Core.Texl.Intellisense.SignatureHelp.SignatureHelp",
                 "Microsoft.PowerFx.Core.Texl.Intellisense.SignatureHelp.SignatureInformation",
-                "Microsoft.PowerFx.Core.ITypeMarshallerProvider",
-                "Microsoft.PowerFx.Core.ITypeMarshaller",
-                "Microsoft.PowerFx.Core.ObjectMarshallerProvider",
-                "Microsoft.PowerFx.Core.ObjectMarshaler",
-                "Microsoft.PowerFx.Core.PrimitiveMarshallerProvider",
-                "Microsoft.PowerFx.Core.PrimitiveTypeMarshaler",
-                "Microsoft.PowerFx.Core.TableMarshallerProvider",
-                "Microsoft.PowerFx.Core.TypeMarshallerCache",
-                "Microsoft.PowerFx.Core.ObjectRecordValue",
+                "Microsoft.PowerFx.Core.BuiltinFormulaTypeConversions",
                 "Microsoft.PowerFx.Core.PowerFxConfig",
                 "Microsoft.PowerFx.Core.Public.CheckResult",
                 "Microsoft.PowerFx.Core.Public.ErrorKind",
